requirements.txt
Django>=4.2
djangorestframework>=3.14
langchain
langchain-openai
langchain-community
langchain-experimental
redis
